backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
boto3==1.40.9
botocore==1.40.9
certifi==2025.8.3
cffi==1.17.1
customtkinter==5.2.2
darkdetect==0.8.0
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework-csv==3.0.2
h11==0.16.0
idna==3.10
jmespath==1.0.1
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==10.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-barcode==0.16.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2025.2
s3transfer==0.13.1
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.30.0
trio-websocket==0.12.2
ttkbootstrap==1.14.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
